Move classes to program
