--- universal-application-tool-0.0.1/app/services/applicant/predicate/ConcretePredicateExpressionNode.java
@@ -1,10 +1,0 @@
-package services.applicant.predicate;
-
-public interface ConcretePredicateExpressionNode {
-
-  /**
-   * Returns the type of this node, as a {@link
-   * services.applicant.predicate.PredicateExpressionNodeType}.
-   */
-  PredicateExpressionNodeType getType();
-}
