+++ universal-application-tool-0.0.1/app/services/program/BlockDefinition.java
@@ -116,8 +116,14 @@ public abstract class BlockDefinition {
   @JsonProperty("hidePredicate")
   public abstract Optional<PredicateDefinition> predicate();
 
-  /** A {@link Predicate} that determines whether this is optional or required. */
+  /**
+   * A {@link Predicate} that determines whether this is optional or required. Note as of
+   * 2021-05-25: We no longer consider blocks to be required or optional - a block is required if
+   * shown. Instead, individual questions can be optional or required. This field is kept for
+   * serialization consistency.
+   */
   @JsonProperty("optionalPredicate")
+  @Deprecated
   public abstract Optional<Predicate> optionalPredicate();
 
   /** The list of {@link ProgramQuestionDefinition}s that make up this block. */
@@ -163,8 +169,10 @@ public abstract class BlockDefinition {
     }
 
     @JsonProperty("optionalPredicate")
+    @Deprecated
     public abstract Builder setOptionalPredicate(Optional<Predicate> optional);
 
+    @Deprecated
     public Builder setOptionalPredicate(Predicate optional) {
       return this.setOptionalPredicate(Optional.of(optional));
     }
