+++ universal-application-tool-0.0.1/app/services/program/predicate/ConcretePredicateExpressionNode.java
@@ -1,0 +1,7 @@
+package services.program.predicate;
+
+public interface ConcretePredicateExpressionNode {
+
+  /** Returns the type of this node, as a {@link PredicateExpressionNodeType}. */
+  PredicateExpressionNodeType getType();
+}
