+++ universal-application-tool-0.0.1/app/services/program/predicate/Operator.java
@@ -1,4 +1,4 @@
-package services.applicant.predicate;
+package services.program.predicate;
 
 /** Represents a JsonPath operator (https://github.com/json-path/JsonPath#filter-operators). */
 public enum Operator {
