+++ universal-application-tool-0.0.1/app/services/program/predicate/PredicateExpressionNodeType.java
@@ -1,4 +1,4 @@
-package services.applicant.predicate;
+package services.program.predicate;
 
 public enum PredicateExpressionNodeType {
   AND,
